requirements.txt
requests==2.31.0
GitPython==3.1.40
pandas==2.2.0
jsonschema==4.21.1
tqdm==4.66.1
python-dotenv==1.0.1
